extract backend list and run settings to separate files
