+++ flutter/lib/backend/list.in
@@ -1,4 +1,4 @@
-part of 'bridge.dart';
+part of 'list.dart';
 
 // Always add new tags before TFLITE_TAG, otherwise your backend will never be used
 const _backendsList = [
