+++ flutter/lib/benchmark/benchmark.dart
@@ -11,11 +11,13 @@ import 'package:collection/collection.dart';
 import 'package:device_info/device_info.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:ios_utsname_ext/extension.dart';
+import 'package:mlcommons_ios_app/backend/run_settings.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:wakelock/wakelock.dart';
 
 import 'package:mlcommons_ios_app/backend/bridge.dart';
+import 'package:mlcommons_ios_app/backend/list.dart';
 import 'package:mlcommons_ios_app/backend/native.dart' as native_backend;
 import 'package:mlcommons_ios_app/benchmark/benchmark_result.dart';
 import 'package:mlcommons_ios_app/benchmark/resource_manager.dart';
